updated to mc 1.20

+++ src/main/java/net/voidz/dimension/VoidPlacementHandler.java
@@ -33,7 +33,7 @@ public class VoidPlacementHandler {
             BlockState platformBlock = BlockInit.VOID_BLOCK.getDefaultState();
             for (float u = 0.0F; u < Math.PI * 2; u += (float) Math.PI / 256F) {
                 for (int i = 0; i < 40; i++) {
-                    world.setBlockState(pos.add(Math.sin(u) * i, 0, Math.cos(u) * i), platformBlock);
+                    world.setBlockState(pos.add((int) Math.sin(u) * i, 0, (int) Math.cos(u) * i), platformBlock);
                 }
             }
             // Pretty good centered
